Reduced duplication within Spring Boot application config (#210)
