+++ forgerock-openbanking-aspsp/forgerock-openbanking-aspsp-rs/forgerock-openbanking-rs-api/pom.xml
@@ -36,6 +36,10 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.forgerock.openbanking.common</groupId>
+            <artifactId>forgerock-openbanking-common</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.forgerock.openbanking.aspsp</groupId>
             <artifactId>forgerock-openbanking-uk-aspsp-rs-gateway-server</artifactId>
         </dependency>
