+++ forgerock-openbanking-aspsp/forgerock-openbanking-aspsp-rs/forgerock-openbanking-rs-ui/pom.xml
@@ -40,6 +40,10 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>com.forgerock.openbanking.common</groupId>
+			<artifactId>forgerock-openbanking-common</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>com.forgerock.openbanking.aspsp</groupId>
 			<artifactId>forgerock-openbanking-uk-aspsp-rs-mock-portal-server</artifactId>
 		</dependency>
