+++ forgerock-openbanking-backstage/forgerock-openbanking-metrics/forgerock-openbanking-metrics-services/pom.xml
@@ -40,6 +40,10 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>com.forgerock.openbanking.common</groupId>
+			<artifactId>forgerock-openbanking-common</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>com.forgerock.openbanking.analytics</groupId>
 			<artifactId>forgerock-openbanking-analytics-core</artifactId>
 		</dependency>
