+++ forgerock-openbanking-devportal/forgerock-openbanking-docs/pom.xml
@@ -43,7 +43,10 @@
 	</properties>
 
 	<dependencies>
-
+		<dependency>
+			<groupId>com.forgerock.openbanking.common</groupId>
+			<artifactId>forgerock-openbanking-common</artifactId>
+		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-actuator</artifactId>
