+++ forgerock-openbanking-devportal/forgerock-openbanking-register/pom.xml
@@ -44,6 +44,10 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>com.forgerock.openbanking.common</groupId>
+			<artifactId>forgerock-openbanking-common</artifactId>
+		</dependency>
+		<dependency>
 			<groupId>com.forgerock.openbanking.aspsp</groupId>
 			<artifactId>forgerock-openbanking-uk-aspsp-as-manual-onboarding-server</artifactId>
 		</dependency>
