+++ forgerock-openbanking-jwkms/pom.xml
@@ -35,17 +35,21 @@
     </parent>
 
     <dependencies>
+        <dependency>
+            <groupId>com.forgerock.openbanking.common</groupId>
+            <artifactId>forgerock-openbanking-common</artifactId>
+        </dependency>
 		<dependency>
 			<groupId>dev.openbanking4.spring.security</groupId>
 			<artifactId>spring-security-multi-auth-starter</artifactId>
 		</dependency>
         <dependency>
-            <groupId>com.forgerock.openbanking.clients</groupId>
-            <artifactId>forgerock-openbanking-analytics-webclient</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.forgerock.openbanking.jwkms</groupId>
             <artifactId>forgerock-openbanking-jwkms-server</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.forgerock.openbanking.clients</groupId>
+            <artifactId>forgerock-openbanking-analytics-webclient</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
