+++ pom.xml
@@ -56,6 +56,7 @@
     </properties>
 
     <modules>
+        <module>forgerock-openbanking-common</module>
         <module>forgerock-openbanking-config</module>
         <module>forgerock-openbanking-aspsp</module>
         <module>forgerock-openbanking-jwkms</module>
@@ -68,6 +69,11 @@
     <dependencyManagement>
         <dependencies>
             <!-- OpenBanking -->
+            <dependency>
+                <groupId>com.forgerock.openbanking.common</groupId>
+                <artifactId>forgerock-openbanking-common</artifactId>
+                <version>${project.version}</version>
+            </dependency>
             <dependency>
                 <groupId>com.forgerock.openbanking.analytics</groupId>
                 <artifactId>forgerock-openbanking-analytics-core</artifactId>
@@ -97,6 +103,11 @@
                 <groupId>com.forgerock.openbanking.directory</groupId>
                 <artifactId>forgerock-openbanking-directory-server</artifactId>
                 <version>${ob-directory.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.forgerock.openbanking.aspsp</groupId>
+                <artifactId>forgerock-openbanking-uk-aspsp-common</artifactId>
+                <version>${ob-aspsp.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.forgerock.openbanking.aspsp</groupId>
